DEV-873: get firebird working locally with catalog
Adds a FIREBIRD_HOME environment variable (to be included via
docker-compose.yml) to enable loading firebird assets locally from an
arbitrary location rather than relying on DOCUMENT_ROOT (which may not
be set as expected in development) or falling back on loading assets via
netlify.

Sets clear_env = no to allow passing through env vars from
docker-compose - we might not want to do that if/when we deploy the
catalog in Kubernetes, but it's safe enough in development.