DEV-1170: add title to links in resource access container
Addresses #46; Deque issue 1727511

Note that it is possible that we could have two results with the same
title, in which case two things would still end up with the same link
text, although that would be the case for the context available to a
sighted user too.

I am only including the first title in the link text for screen readers
rather than additional titles; I don't have an example of an item with
multiple titles handy to get a sense of whether that would be necessary
to include (I suspect it would be less than helpful given that titles
can be rather long anyway)

Note also that I am not including the title in original script here as I'm
assuming we can't tag the language appropriately for the screen reader
anyay. (although there is a way in MARC to identify what script the 880
uses, it's somewhat limited and it's unclear whether many of our records
use it)